move core tests to separate file
